requirements.txt
Flask
Flask-Reuploaded
Flask-Login
Flask-JSON
werkzeug
Pillow
requests
bs4
lxml
html5lib
boto3